test(app): move test network/node utils to utils module
